+++ lib/flutter_better_auth.dart
@@ -61,7 +61,6 @@ export 'core/models/session/session.dart';
 
 export 'core/models/user/user.dart';
 export 'core/models/verification/verification.dart';
-export 'plugins/phone/models/sign_in_phone_body.dart';
 export 'presentation/better_auth_consumer.dart';
 export 'presentation/better_auth_inherit.dart';
 export 'presentation/better_auth_provider.dart';